Chore: extract common part of indent
